pyproject.toml
[project]
name = "airflow-dbt-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow-providers-google==10.24.0",
    "google-cloud-storage>=1.32.0,<3.0.0",
    "pandas>=2.1.2,<2.2.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.0,<3.0",
    "pytest>=8.0,<9.0",
    "python-dotenv>=0.21.0,<0.22.0",
    "slack-sdk>=3.35.0,<4.0",
    "sqlalchemy>=1.4.49,<2.0",
    "taskipy>=1.14.1",
]


[tool.pytest.ini_options]
pythonpath = ["."]

[tool.taskipy.tasks]
run = "docker-compose up -d"
kill = "docker-compose down -v"
test = "pytest tests/test_validator_raw.py -v"
test_all = "pytest tests/ -v"
